src/main/java/com/vermouth/sell/service/impl/CategoryServiceImpl.java
package com.vermouth.sell.service.impl;

import com.vermouth.sell.dataobject.ProductCategory;
import com.vermouth.sell.repository.ProductCategoryRepository;
import com.vermouth.sell.service.CategoryService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Set;

@Service
public class CategoryServiceImpl implements CategoryService {
    @Autowired
    private ProductCategoryRepository repository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return repository.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return repository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(Set<Integer> categoryTypeSet) {
        return repository.findByCategoryTypeIn(categoryTypeSet);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return repository.save(productCategory);
    }
}
